Remove support for mattermost 5
